Initial Commit
This is a copy of an existing repository that has unfortunately been
poisoned by secret data. Rather than filtering that data out, it is
simpler to start this github repo fresh with a "clean" commit.

With the exception of merge commits, I am the sole author of all of this
code (and therefor responsible for all pythonic crimes that may be
present in this commmit).